Remove flight-control api
