+++ backend/src/index.ts
@@ -2,7 +2,6 @@ import express from 'express';
 import cors from 'cors';
 import timeslotRouter from './routes/timeslots';
 import reservationRouter from './routes/reservations';
-import flightControl from './routes/flightcontrol';
 
 const app = express();
 
@@ -13,6 +12,5 @@ app.get('/api', async (_req: any, res: express.Response) => {
 });
 app.use('/api/timeslots', timeslotRouter);
 app.use('/api/reservations', reservationRouter);
-app.use('/api/flight-control', flightControl);
 
 export default app;
