+++ backend/src/services/reservationService.ts
@@ -1,8 +1,8 @@
 import { Op } from 'sequelize';
-import { ReservationEntry } from '@lentovaraukset/shared/src';
+import { ReservationEntry, ReservationStatus } from '@lentovaraukset/shared/src';
 import countMostConcurrent from '@lentovaraukset/shared/src/overlap';
 import timeslotService from '@lentovaraukset/backend/src/services/timeslotService';
-import { Reservation } from '../models';
+import { Reservation, Timeslot, User } from '../models';
 
 const maxConcurrentReservations: number = 3;
 
@@ -88,10 +88,44 @@ const updateById = async (
   }
 };
 
+const getReservationStatus = async (): Promise<ReservationStatus> => {
+  const returnedTimeSlots = await Timeslot.findAll({
+    include: {
+      model: Reservation,
+      attributes: ['start', 'end', 'info'],
+    },
+  });
+  const availableSlots = returnedTimeSlots.map(({
+    id, start, end, ...rest
+  }) => ({
+    id,
+    start,
+    end,
+    // any because sequelize typing only does basic values, not relations
+    freeSlotsAmount: (rest as any).maxAmount - (rest as any).reservations.length,
+  }));
+  const reservedSlots = await Reservation.findAll({
+    include: {
+      model: User,
+      attributes: ['name', 'role'],
+    },
+  });
+
+  // we don't have users yet
+  const reservationEntries: ReservationEntry[] = reservedSlots.map(({
+    id, start, end, aircraftId, info,
+  }) => ({
+    id, start, end, aircraftId, info, user: 'NYI', phone: 'NYI',
+  }));
+
+  return { availableSlots, reservedSlots: reservationEntries };
+};
+
 export default {
   createReservation,
   getInTimeRange,
   deleteById,
   updateById,
   getReservationFromRange,
+  getReservationStatus,
 };
