+++ frontend/src/queries/reservations.ts
@@ -1,4 +1,10 @@
 import { EventInput } from '@fullcalendar/core';
+import { ReservationStatus } from '@lentovaraukset/shared/src';
+
+const getAllReservations = async (): Promise<ReservationStatus> => {
+  const res = await fetch(`${process.env.BASE_PATH}/api/reservations/status`);
+  return res.json();
+};
 
 const getReservations = async (from: Date, until: Date): Promise<EventInput[]> => {
   const res = await fetch(`${process.env.BASE_PATH}/api/reservations?from=${from.toISOString()}&until=${until.toISOString()}`);
@@ -49,4 +55,5 @@ export {
   addReservation,
   modifyReservation,
   deleteReservation,
+  getAllReservations,
 };
